requirements.txt
attrs
beautifulsoup4
flask
requests
